requirements.txt
ffmpeg-python
numpy
matplotlib
scipy